Move DimensionType to type_enums dir
`DimensionType` is another extended enum used to restrict typing. As
its neither a protocol or an implementation, but used in both, it's
been moved to the `type_enums` directory.
