+++ dbt_semantic_interfaces/dbt_semantic_interfaces/objects/elements/dimension.py
@@ -5,21 +5,10 @@ from typing import Optional
 from dbt_semantic_interfaces.objects.base import HashableBaseModel, ModelWithMetadataParsing
 from dbt_semantic_interfaces.objects.metadata import Metadata
 from dbt_semantic_interfaces.references import DimensionReference, TimeDimensionReference
-from dbt_semantic_interfaces.enum_extension import ExtendedEnum
+from dbt_semantic_interfaces.type_enums.dimension_type import DimensionType
 from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 ISO8601_FMT = "YYYY-MM-DD"
-
-
-class DimensionType(ExtendedEnum):
-    """Determines types of values expected of dimensions."""
-
-    CATEGORICAL = "categorical"
-    TIME = "time"
-
-    def is_time_type(self) -> bool:
-        """Checks if this type of dimension is a time type"""
-        return self in [DimensionType.TIME]
 
 
 class DimensionValidityParams(HashableBaseModel):
