+++ metricflow/model/spec_converters.py
@@ -13,7 +13,7 @@ from typing import List
 
 from metricflow.dataset.dataset import DataSet
 from dbt_semantic_interfaces.objects.constraints.where import WhereClauseConstraint
-from dbt_semantic_interfaces.objects.elements.dimension import DimensionType
+from dbt_semantic_interfaces.type_enums.dimension_type import DimensionType
 from dbt_semantic_interfaces.objects.elements.measure import Measure
 from metricflow.protocols.semantics import DataSourceSemanticsAccessor
 from metricflow.query.query_exceptions import InvalidQueryException
