+++ metricflow/model/validations/agg_time_dimension.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from dbt_semantic_interfaces.objects.data_source import DataSource
-from dbt_semantic_interfaces.objects.elements.dimension import DimensionType
+from dbt_semantic_interfaces.type_enums.dimension_type import DimensionType
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from dbt_semantic_interfaces.references import DataSourceElementReference, TimeDimensionReference
 from metricflow.model.validations.validator_helpers import (
