+++ metricflow/model/validations/measures.py
@@ -4,7 +4,7 @@ from typing import DefaultDict, Dict, List, Set
 from more_itertools import bucket
 
 from dbt_semantic_interfaces.objects.aggregation_type import AggregationType
-from dbt_semantic_interfaces.objects.elements.dimension import DimensionType
+from dbt_semantic_interfaces.type_enums.dimension_type import DimensionType
 from dbt_semantic_interfaces.objects.metric import Metric
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from dbt_semantic_interfaces.references import MetricModelReference, MeasureReference
