+++ metricflow/test/model/validations/test_agg_time_dimension.py
@@ -1,7 +1,7 @@
 import pytest
 
 from metricflow.model.model_validator import ModelValidator
-from dbt_semantic_interfaces.objects.elements.dimension import DimensionType
+from dbt_semantic_interfaces.type_enums.dimension_type import DimensionType
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from metricflow.model.validations.agg_time_dimension import AggregationTimeDimensionRule
 from metricflow.model.validations.validator_helpers import ModelValidationException
