+++ metricflow/test/model/validations/test_data_sources.py
@@ -1,9 +1,10 @@
 import pytest
 
 from dbt_semantic_interfaces.objects.data_source import MutabilityType, Mutability
-from dbt_semantic_interfaces.objects.elements.dimension import Dimension, DimensionType, DimensionTypeParams
+from dbt_semantic_interfaces.objects.elements.dimension import Dimension, DimensionTypeParams
 from metricflow.model.validations.validator_helpers import ModelValidationException
 from metricflow.test.model.validations.helpers import data_source_with_guaranteed_meta
+from dbt_semantic_interfaces.type_enums.dimension_type import DimensionType
 from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
